+++ app/bot/__init__.py
@@ -1,0 +1,6 @@
+"""
+The telegram bot.
+
+It shows list on available bikes in the store.
+Also, user can be informed about unavailable bike family after subscription, when it will be in the store again.
+"""
