+++ app/bot/common_handlers.py
@@ -1,19 +1,12 @@
-"""
-This is the canyon new bike's bot.
+"""List of common bot handlers."""
 
-It answers to any incoming text messages with the list of all commands.
-"""
-import logging
 from itertools import groupby
 from typing import Generator, List
 
-from aiogram import Bot, Dispatcher, executor, types
-from aiogram.contrib.fsm_storage.memory import MemoryStorage
+from aiogram import types
 from aiogram.utils.markdown import hlink
 
-from app import bot_subscription
 from app.bike_model import Bike, CatalogFamily
-from app.settings import app_settings
 from app.storage import get_catalog
 
 
@@ -62,31 +55,3 @@ async def show_catalog(message: types.Message) -> None:  # noqa: WPS210
             parse_mode='HTML',
             disable_web_page_preview=True,
         )
-
-
-def main() -> None:
-    """Telegram bot app runner."""
-    bot = Bot(token=app_settings.bot_token)
-    storage = MemoryStorage()
-
-    router = Dispatcher(bot, storage=storage)
-    router.register_message_handler(send_welcome, commands=['start', 'help'])
-    router.register_message_handler(show_catalog, commands=['catalog'])
-    router.register_message_handler(bot_subscription.start_subscription, commands=['subscribe'])
-    router.register_message_handler(bot_subscription.cancel_subscription, state='*', commands=['cancel'])
-    router.register_message_handler(
-        bot_subscription.process_subscription,
-        state=bot_subscription.CreateSubscription.family_name,
-    )
-    router.register_message_handler(bot_subscription.remove_subscription, commands=['unsubscribe'])
-    router.register_message_handler(bot_subscription.show_subscriptions, commands=['subscriptions_list'])
-    executor.start_polling(router, skip_updates=True)
-
-
-if __name__ == '__main__':
-    logging.basicConfig(
-        level=logging.DEBUG if app_settings.debug else logging.INFO,
-        format='%(asctime)s %(levelname)-8s %(message)s',  # noqa: WPS323
-    )
-
-    main()
